Fix build break.
Move the new Add/Remove certificate methods to a new Interop file, since
they're more about acting on a cert store than the keychain.
